se agregaron todos los required en angular

+++ FylsaApp/src/app/Associates/associates-remove/associates-remove.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import {AssociatesService} from '../associates.service';
 import {Associates} from '../../Models/associates';
 
@@ -10,16 +11,22 @@ import {Associates} from '../../Models/associates';
 })
 export class AssociatesRemoveComponent implements OnInit {
 
-     associates: Associates;    
+    deleteForm: FormGroup; // Declare the signupForm
+    idCtrl: FormControl;
+    associates: Associates;
 
-  constructor(private  associatesService: AssociatesService) {
+  constructor(private  associatesService: AssociatesService, private fb:FormBuilder) {
       this.associates = new Associates("","","",0);
-      
+      this.idCtrl = new FormControl('', Validators.required);
    }
   ngOnInit() {
+    this.deleteForm  = this.fb.group({
+      'inputId': this.idCtrl
+    });
   }
  deleteAssociate(){
+      if(this.deleteForm.valid) {
          this.associatesService.deleteAssociate(this.associates._id).subscribe((res) => alert(JSON.stringify(res["success"])));
-    }
+       }
+  }
 }
-
