+++ FylsaApp/src/app/Company/company-delete/company-delete.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import {CompanyService} from '../company.service';
 import {Company} from '../../Models/company';
 
@@ -9,16 +10,25 @@ import {Company} from '../../Models/company';
 })
 export class CompanyDeleteComponent implements OnInit {
 
-  company: Company;
-    constructor(private companyService:CompanyService) {
+    deleteForm: FormGroup; // Declare the signupForm
+    idCtrl: FormControl;
+    company: Company;
+    
+    constructor(private companyService:CompanyService, private fb:FormBuilder) {
      this.company = new Company("","","","");
+     this.idCtrl = new FormControl('', Validators.required);
    }
 
   ngOnInit() {
+    this.deleteForm  = this.fb.group({
+      'inputId': this.idCtrl
+    });
   }
 
      deleteCompany(){
+       if(this.deleteForm.valid) {
         console.log(this.company._id);
          this.companyService.deleteCompany(this.company._id).subscribe((res) => alert(JSON.stringify(res["success"])));
+       }
     }
 }
